// ui/src/generated/api/model/dtoAccountResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoAccountResponseRole } from "./dtoAccountResponseRole";

export interface DtoAccountResponse {
  createdAt: string;
  disabled: boolean;
  disabledAt?: string;
  disabledReason?: string;
  email: string;
  id: string;
  rating: number;
  role: DtoAccountResponseRole;
  solvedCount: number;
  submissionCount: number;
  username: string;
}